+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 * Update README
 * add Color
 * remove unused file
+
+# 0.0.4
+* Update CI
+* export files
+* add Button Components
+* add Textfield
+* add DropDown Button
+* add Icon